refactor the Excel PL-parsers into separate file
